Added the minimal compiling example.
Here is how it works.
You have to deploy the war in the tomcat webapps/ directory, after
that new directory in the webapps/ will appear, go there and see
the compiled classes. How to access the websocket. You cannot simply
go to the link in the browser, you have to use some smart wrapper
that would send the initial request for handshake to the server
endpoint. This can be done, for example, by the web developer's
console in Firefox, but will be done from another Java client in
this laboratory work.

Another interesting thing. The webapps directory is used for the
web artifacts, that means that even if I tried to develop a client
in JavaScript, I could do it in the webapp/ directory and that
content would be served to the user.

And the next thing. Since Servlet 3.0, web.xml descriptors are no longer
needed. It is now the job of the container (like Tomcat or Jetty) to
look for stuff in the classes and determine the endpoints from them.
Currently Tomcat is version 8 (9 is being developed), and I do not need
any of those configs at all, especially doing simple websocket
development.